requirements.txt
TextBlob
pickle5
scikit-learn
matplotlib
googleapiclient.discovery
sklearn.feature_extraction.text